+++ mathics/algorithm/parts.py
@@ -7,7 +7,7 @@ Algorithms to access and manipulate elements in nested lists / expressions
 
 from mathics.core.expression import Expression
 from mathics.core.symbols import Symbol
-from mathics.core.atoms import Integer, from_python
+from mathics.core.atoms import Integer, Integer1, from_python
 from mathics.core.systemsymbols import SymbolInfinity
 
 from mathics.builtin.exceptions import (
@@ -341,7 +341,7 @@ def python_levelspec(levelspec):
     def value_to_level(expr):
         value = expr.get_int_value()
         if value is None:
-            if expr == Expression("DirectedInfinity", 1):
+            if expr == Expression("DirectedInfinity", Integer1):
                 return None
             else:
                 raise InvalidLevelspecError