+++ mathics/builtin/colors/color_directives.py
@@ -21,6 +21,7 @@ from mathics.core.atoms import (
     Real,
     String,
     from_python,
+    MachineReal,
 )
 
 from mathics.core.symbols import Symbol, SymbolList
@@ -192,7 +193,7 @@ class _Color(_GraphicsElement):
         return self.to_rgba()
 
     def to_expr(self):
-        return Expression(self.get_name(), *self.components)
+        return Expression(self.get_name(), *[MachineReal(c) for c in self.components])
 
     def to_rgba(self):
         return self.to_color_space("RGB")
@@ -586,5 +587,5 @@ def color_to_expression(components, colorspace):
         converted_color_name = "Hue"
     else:
         converted_color_name = colorspace + "Color"
-
+    print(components)
     return Expression(converted_color_name, *components)