+++ mathics/builtin/intfns/divlike.py
@@ -10,8 +10,8 @@ from itertools import combinations
 from mathics.version import __version__  # noqa used in loading to check consistency.
 from mathics.builtin.base import Builtin, Test, SympyFunction
 from mathics.core.expression import Expression
-from mathics.core.atoms import Integer
-from mathics.core.symbols import Symbol
+from mathics.core.atoms import Integer, from_python
+from mathics.core.symbols import Symbol, SymbolList
 
 
 class CoprimeQ(Builtin):
@@ -372,6 +372,8 @@ class QuotientRemainder(Builtin):
                 return evaluation.message(
                     "QuotientRemainder", "divz", Expression("QuotientRemainder", m, n)
                 )
-            return Expression("List", Integer(py_m // py_n), (py_m % py_n))
+            return Expression(
+                SymbolList, Integer(py_m // py_n), from_python(py_m % py_n)
+            )
         else:
             return Expression("QuotientRemainder", m, n)