+++ mathics/builtin/lists.py
@@ -334,7 +334,7 @@ class Delete(Builtin):
         try:
             return delete_one(expr, pos)
         except PartRangeError:
-            evaluation.message("Part", "partw", Expression(SymbolList, pos), expr)
+            evaluation.message("Part", "partw", Expression(SymbolList, position), expr)
 
     def apply(self, expr, positions, evaluation):
         "Delete[expr_, positions___]"
@@ -2082,6 +2082,7 @@ class _Cluster(Builtin):
                 SymbolList, *[Expression(SymbolList, *c) for c in clusters]
             )
         elif mode == "components":
+            clusters = [from_python(c) for c in clusters]
             return Expression(SymbolList, *clusters)
         else:
             raise ValueError("illegal mode %s" % mode)