+++ mathics/builtin/strings.py
@@ -33,6 +33,7 @@ from mathics.core.atoms import (
     Integer,
     Integer0,
     Integer1,
+    from_python,
 )
 
 from mathics.core.parser import MathicsFileLineFeeder, parse
@@ -567,7 +568,7 @@ class LetterNumber(Builtin):
                     letter_number(c, start_ord)[0] if c.isalpha() else 0
                     for c in py_chars
                 ]
-                return Expression(SymbolList, *r)
+                return Expression(SymbolList, *[from_python(rr) for rr in r])
         elif chars.has_form("List", 1, None):
             result = []
             for leaf in chars.leaves: